+++ db.json
@@ -721,6 +721,22 @@
       "password": "0000",
       "content": "김치가 너무 맛있어요.",
       "date": "2022-12-12 12:13:35"
+    },
+    {
+      "id": 41,
+      "kimchiId": 40,
+      "nickname": "김치짱",
+      "password": "0000",
+      "content": "김치가 너무 맛있어요.",
+      "date": "2022-12-12 12:13:35"
+    },
+    {
+      "id": 42,
+      "kimchiId": 40,
+      "nickname": "김치포에버",
+      "password": "0000",
+      "content": "김치가 너무 맛있어요.",
+      "date": "2022-12-12 12:13:35"
     }
   ],
   "cart": [
